moved test to test_california_housing.py
